requirements.txt
nltk
pyyaml